pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gvclass"
version = "1.1.0"
description = "Giant Virus Classification Pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "NeLLi Team"},
]
maintainers = [
    {name = "NeLLi Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "pandas>=2.0",
    "numpy>=2.3",
    "biopython>=1.85",
    "pyhmmer>=0.10.0", 
    "pytrimal>=0.8.0",
    "pyfamsa>=0.5.0",
    "pyrodigal>=3.0.0",
    "pyswrd>=0.1.0",
    "veryfasttree>=3.0.0",
    "dask>=2024.7.0",
    "distributed>=2024.7.0",
    "dask-jobqueue>=0.9.0",
    "prefect>=2.21.0",
    "prefect-dask>=0.3.0",
    "networkx>=3.0",
    "cloudpickle>=3.0",
]

[project.urls]
Homepage = "https://github.com/NeLLi-team/gvclass"
Documentation = "https://github.com/NeLLi-team/gvclass/wiki"
Repository = "https://github.com/NeLLi-team/gvclass"
Issues = "https://github.com/NeLLi-team/gvclass/issues"

[project.scripts]
gvclass = "src.bin.gvclass:cli"
gvclass-dask = "src.bin.gvclass_cli:cli"
gvclass-prefect = "src.bin.gvclass_prefect:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.json"]